Fixed the bug for the Investment app
Fix the path issue in the financial app.
